Move Timeline to shared UI
