+++ ui/packages/components/package.json
@@ -11,7 +11,9 @@
   "dependencies": {
     "@headlessui/tailwindcss": "0.1.2",
     "@monaco-editor/react": "4.5.2",
+    "@radix-ui/react-accordion": "1.1.2",
     "@radix-ui/react-tooltip": "1.0.6",
+    "framer-motion": "10.16.4",
     "monaco-editor": "0.34.1",
     "next": "13.4.12",
     "react": "18.2.0",
