+++ ui/packages/components/src/OutputCard/OutputCard.tsx
@@ -7,7 +7,7 @@ interface OutputCardProps {
   content: string;
 }
 
-export default function OutputCard({ type, content }: OutputCardProps) {
+export function OutputCard({ type, content }: OutputCardProps) {
   let { message, errorName, output } = renderOutput({ type, content });
 
   if (!message && !output) return null;
