+++ ui/packages/components/src/OutputCard/index.ts
@@ -1,0 +1,1 @@
+export { OutputCard } from './OutputCard';
