+++ ui/packages/components/src/Timeline/TimelineNode/TimelineNodeHeader/index.ts
@@ -1,0 +1,1 @@
+export { TimelineNodeHeader } from './TimelineNodeHeader';
